rename statevector_kernel file
